Added fasterizeDT() function, a wrapper for fasterize::fasterize()
As described in an Issue filed at the **fasterize** package's GitHub
repository (https://github.com/ecohealthalliance/fasterize/issues/26),
`fasterize` returns rasters that match the class of their template
raster, not of the field that the user selected to be burned into
them. In addition, when passed a character or a factor field, it does
not -- as would often be desireable -- return a factor-valued raster.

`fasterizeDT()` addresses both of these issues, consistently returning
a raster whose class (and, if a factor, levels) matches that of the
user-selected field in the input polygon object.

In addition, `fasterizeDT()` accepts `sp::SpatialPolygonsDataFrame()`
objects, converting them to `sf::sf()` objects before passing them on
to `fasterize::fasterize()`.