Made subsDT() respect NA and NaN as replacement values from dict
As kindly pointed out and well elucidated by Michael Koontz in [this
Issue](https://github.com/JoshOBrien/rasterDT/issues/1)and [this
PR](https://github.com/JoshOBrien/rasterDT/pull/2), the output of
`rasterDT::subsDT()` did not match that of `raster::subs()` when
`subWithNA=FALSE` and one or more of the keys in `dict=` mapped values
to `NA` or `NaN`.

The fix I've applied here is perhaps more opaque than the one that
Michael proposed, but has the advantage of only coming into play when
`subsWithNA=FALSE` and when the user-supplied dictionary does map some
values to `NA` or `NaN`. `microbenchmark::microbenchmark()` has given
me somewhat erratic results when investigating which approach is
fastest, but at least in my tests, this approach seems to perform well.

It's probably worth noting that, for rasters too large to process in
memory (but not for those that can be so processed),
`rasterDT::subsDT()` with `subsWithNA=FALSE` will convert values that
should be mapped to `NaN` to `NA` instead. `raster::subs()` has this
same infelicitous behavior. In both cases, the fault seems to be with
the `raster::writeValues()` step, which will take in a supplied
argument to `vals=` containing `NaN`s and write out a raster that has
converted all of them to `NA`s. `raster::writeRaster()` does not seem
to have that same issue, so this may be a bug that could be fixed by
comparing its code with that of `raster::writeValues()`.