Removed backticks to avoid CRAN check complaint abt \textasciigrave
I got a new LaTeX related ERROR when running `R CMD check --as-cran`,
and the one Google hit I got for it was this recent (10 day old) post
by Spencer Graves, who was experiencing the same problem.
https://stat.ethz.ch/pipermail/r-package-devel/2020q1/005051.html.

Unlike him, it was a small matter to remove the few backticks in
my *.R files (all of them in comments as delimiters of code snippets).